bugfix - use lambdas as middleware
If a child controller overrides the render method, lambdas is no longer available unless super is called (which is often not desired). Due to the nature of lambdas it should have been implemented as a middleware, as per template mixins.

* added next param to lambdas
* call next after res.locals extended with new lambda
* remove render method from base-controller
* apply middleware with controller.use interface
* altered lambdas to return a function so options can potentially be passed in in the future
