+++ lib/base-controller.js
@@ -5,12 +5,13 @@ const i18nLookup = require('i18n-lookup');
 const Mustache = require('mustache');
 const helpers = require('./util/helpers');
 const Controller = require('hof-form-wizard').Controller;
-const lambdas = require('./mixins/lambdas');
+const lambdas = require('./middleware/lambdas');
 
 module.exports = class BaseController extends Controller {
   constructor(options) {
     super(options);
     this.confirmStep = options.confirmStep || '/confirm';
+    this.use(lambdas());
   }
 
   get(req, res, callback) {
@@ -60,11 +61,6 @@ module.exports = class BaseController extends Controller {
     }
 
     return next;
-  }
-
-  render(req, res) {
-    lambdas(req, res);
-    super.render(req, res);
   }
 
   getBackLink(req, res) {
