+++ lib/middleware/lambdas.js
@@ -1,6 +1,6 @@
 'use strict';
 
-module.exports = function lambdas(req, res) {
+module.exports = () => (req, res, next) => {
   Object.assign(res.locals, {
     renderField() {
       return function renderFieldMixin() {
@@ -20,4 +20,5 @@ module.exports = function lambdas(req, res) {
       };
     }
   });
+  next();
 };
